BVH code uses floats instead of doubles..

+++ opengl_acquisition_shared_library/opengl_depth_and_color_renderer/src/Library/MotionCaptureLoader/bvh_loader.c
@@ -574,10 +574,10 @@ int bvh_changeJointDimensions(
        bvhMotion->jointHierarchy[jID].offset[angleZ] *= zScale;
        //fprintf(stderr,"offset is now %0.2f %0.2f %0.2f\n",bvhMotion->jointHierarchy[jID].offset[0],bvhMotion->jointHierarchy[jID].offset[1],bvhMotion->jointHierarchy[jID].offset[2]);
 
-       double * m = bvhMotion->jointHierarchy[jID].staticTransformation;
-       m[0] =1.0;  m[1] =0.0;  m[2] =0.0;  m[3] = (double) bvhMotion->jointHierarchy[jID].offset[0];
-       m[4] =0.0;  m[5] =1.0;  m[6] =0.0;  m[7] = (double) bvhMotion->jointHierarchy[jID].offset[1];
-       m[8] =0.0;  m[9] =0.0;  m[10]=1.0;  m[11]= (double) bvhMotion->jointHierarchy[jID].offset[2];
+       float * m = bvhMotion->jointHierarchy[jID].staticTransformation;
+       m[0] =1.0;  m[1] =0.0;  m[2] =0.0;  m[3] = (float) bvhMotion->jointHierarchy[jID].offset[0];
+       m[4] =0.0;  m[5] =1.0;  m[6] =0.0;  m[7] = (float) bvhMotion->jointHierarchy[jID].offset[1];
+       m[8] =0.0;  m[9] =0.0;  m[10]=1.0;  m[11]= (float) bvhMotion->jointHierarchy[jID].offset[2];
        m[12]=0.0;  m[13]=0.0;  m[14]=0.0;  m[15]=1.0;
 
       return 1;
@@ -622,10 +622,10 @@ int bvh_scaleAllOffsets(
       bvhMotion->jointHierarchy[jID].offset[angleY] = bvhMotion->jointHierarchy[jID].offset[1] * scalingRatio;
       bvhMotion->jointHierarchy[jID].offset[angleZ] = bvhMotion->jointHierarchy[jID].offset[2] * scalingRatio;
 
-       double * m = bvhMotion->jointHierarchy[jID].staticTransformation;
-       m[0] =1.0;  m[1] =0.0;  m[2] =0.0;  m[3] = (double) bvhMotion->jointHierarchy[jID].offset[0];
-       m[4] =0.0;  m[5] =1.0;  m[6] =0.0;  m[7] = (double) bvhMotion->jointHierarchy[jID].offset[1];
-       m[8] =0.0;  m[9] =0.0;  m[10]=1.0;  m[11]= (double) bvhMotion->jointHierarchy[jID].offset[2];
+       float * m = bvhMotion->jointHierarchy[jID].staticTransformation;
+       m[0] =1.0;  m[1] =0.0;  m[2] =0.0;  m[3] = (float) bvhMotion->jointHierarchy[jID].offset[0];
+       m[4] =0.0;  m[5] =1.0;  m[6] =0.0;  m[7] = (float) bvhMotion->jointHierarchy[jID].offset[1];
+       m[8] =0.0;  m[9] =0.0;  m[10]=1.0;  m[11]= (float) bvhMotion->jointHierarchy[jID].offset[2];
        m[12]=0.0;  m[13]=0.0;  m[14]=0.0;  m[15]=1.0;
     }
